+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 3.5.8
+
 - Autumn Eladrin now have correct level spells for Cure Wounds.
 - Custom actions with no descriptions would cause import failure.
-- Link Adept Marksman resources for Gunslinger class.
+- Link Adept Marksman resources for Gunslinger subclass.
 
 # 3.5.7